Update variable binding with constant time lookup
I actually struggled to understand the whole project (domain area is really complicated), so I started rewriting it to C# (the language I'm more convenient with). In the process of rewriting found not a very optimal way to find variable value in another binding. Basically I'll try to play with my own ideas, and Go would really slow down the process, so I had to make the environment more friendly (and to actually understand how the code works, because you know, you understand something way faster when you're not reading but writing it down).